+++ CHANGELOG.md
@@ -71,3 +71,11 @@
 ## ver 2.8.1 Minor Layout Changes in SWDA Dashboard - MARLO BARUA
 
 ## ver 2.9 Added Dashboard Indicator in the Header and Sidebar - MARLO BARUA
+
+## ver 3.0 Added Agencies and Searchbar - MARLO BARUA
+
+## ver 3.1 Fix the Positioning of Doughnut Chart and Added Table Coloumns in the Backend for Agencies - MARLO BARUA
+
+## ver 3.2 Added Agency Details Page in the SWDA Dashboard - MARLO BARUA
+
+## ver 3.3 Agency Details is now Dynamic using Query parameters - MARLO BARUA